requirements.txt
google-api-python-client==2.123.0
python-dotenv==1.0.1
pandas==2.2.2
streamlit==1.34.0